+++ fedify/federation/middleware.ts
@@ -231,7 +231,7 @@ export class FederationImpl<TContextData>
   inboxRetryPolicy: RetryPolicy;
   activityTransformers: readonly ActivityTransformer<TContextData>[];
   tracerProvider: TracerProvider;
-  defaultHttpMessageSignaturesSpec?: HttpMessageSignaturesSpec;
+  firstKnock?: HttpMessageSignaturesSpec;
 
   constructor(options: FederationOptions<TContextData>) {
     super();
@@ -385,7 +385,7 @@ export class FederationImpl<TContextData>
             specDeterminer: new KvSpecDeterminer(
               this.kv,
               this.kvPrefixes.httpMessageSignaturesSpec,
-              options.defaultHttpMessageSignaturesSpec,
+              options.firstKnock,
             ),
             tracerProvider: this.tracerProvider,
           }));
@@ -400,8 +400,7 @@ export class FederationImpl<TContextData>
     this.activityTransformers = options.activityTransformers ??
       getDefaultActivityTransformers<TContextData>();
     this.tracerProvider = options.tracerProvider ?? trace.getTracerProvider();
-    this.defaultHttpMessageSignaturesSpec =
-      options.defaultHttpMessageSignaturesSpec;
+    this.firstKnock = options.firstKnock;
   }
 
   _initializeRouter() {
@@ -647,7 +646,7 @@ export class FederationImpl<TContextData>
         specDeterminer: new KvSpecDeterminer(
           this.kv,
           this.kvPrefixes.httpMessageSignaturesSpec,
-          this.defaultHttpMessageSignaturesSpec,
+          this.firstKnock,
         ),
         tracerProvider: this.tracerProvider,
       });
@@ -1102,7 +1101,7 @@ export class FederationImpl<TContextData>
             specDeterminer: new KvSpecDeterminer(
               this.kv,
               this.kvPrefixes.httpMessageSignaturesSpec,
-              this.defaultHttpMessageSignaturesSpec,
+              this.firstKnock,
             ),
             tracerProvider: this.tracerProvider,
           }),
@@ -2845,7 +2844,7 @@ export class InboxContextImpl<TContextData> extends ContextImpl<TContextData>
             specDeterminer: new KvSpecDeterminer(
               this.federation.kv,
               this.federation.kvPrefixes.httpMessageSignaturesSpec,
-              this.federation.defaultHttpMessageSignaturesSpec,
+              this.federation.firstKnock,
             ),
           }),
         );